Resolve conflicting property name in query getAllReviews
Both the reviews table and the users table have a 'name' property,
which refers to the name of the restaurant and the name of the user,
respectively. This caused an issue while I was testing my query in
node, when one value gets overwritten by another. This is resolved
by setting `users.name as user`. When time allows, consider changing
the name property in the reviews table as restaurant to avoid
overwriting values. This function gets all reviews and user data
associated with one restaurant.