Detect potential bug: should end Jest after the test is done
While my code passes all tests so far, I received a warning:

Jest did not exit one second after the test run has completed.
This usually means that there are asynchronous operations that
weren't stopped in your tests. Consider running Jest with
`--detectOpenHandles` to troubleshoot this issue. This is because
the connection with database is still running. Will fix this issue
later.